chore: changes based on comments
