+++ Sources/DesignSystem/Controllers/Protocols/BaseViewModel.swift
@@ -3,6 +3,7 @@ import Foundation
 @MainActor
 public protocol BaseViewModel {
     var rightBarButtonTitle: GDSLocalisedString? { get }
+    var backButtonTitle: GDSLocalisedString? { get }
     var backButtonIsHidden: Bool { get }
     
     func didAppear()
