+++ Sources/DesignSystem/Controllers/Protocols/GDSErrorScreenViewModel.swift
@@ -1,0 +1,39 @@
+public protocol GDSErrorScreenViewModel: BaseViewModel, GDSScreenViewModel {
+    init(
+        screenStyle: GDSScreenStyle,
+        body: [any ContentViewModel],
+        movableFooter: [any ContentViewModel],
+        footer: [any ContentViewModel],
+        rightBarButtonTitle: GDSLocalisedString?,
+        backButtonTitle: GDSLocalisedString?,
+        backButtonIsHidden: Bool,
+        didAppearAction: (() -> Void)?,
+        didDismissAction: (() -> Void)?
+    )
+}
+
+extension GDSErrorScreenViewModel {
+    public init(
+        icon: GDSImageViewModel,
+        body: [any ContentViewModel],
+        movableFooter: [any ContentViewModel],
+        footer: [any ContentViewModel],
+        rightBarButtonTitle: GDSLocalisedString? = nil,
+        backButtonTitle: GDSLocalisedString? = nil,
+        backButtonIsHidden: Bool = true,
+        didAppearAction: (() -> Void)? = nil,
+        didDismissAction: (() -> Void)? = nil
+    ) {
+        self = Self(
+            screenStyle: .error,
+            body: [icon] + body,
+            movableFooter: movableFooter,
+            footer: footer,
+            rightBarButtonTitle: rightBarButtonTitle,
+            backButtonTitle: backButtonTitle,
+            backButtonIsHidden: backButtonIsHidden,
+            didAppearAction: didAppearAction,
+            didDismissAction: didDismissAction
+        )
+    }
+}
