+++ Sources/DesignSystem/Controllers/Protocols/GDSScreenViewModel.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+@MainActor
+public protocol GDSScreenViewModel {
+    var screenStyle: GDSScreenStyle { get }
+    var body: [any ContentViewModel] { get }
+    var movableFooter: [any ContentViewModel] { get }
+    var footer: [any ContentViewModel] { get }
+}
